Implement Dispatcher
Dispatcher does the job of collecting external code and dispatching it to LiveUpdate.

Any external service can send code to the meteor app for Hot-Loading by "POST"ing the code to `/_hotloader` route. That means we can implement a generic watcher for watching changes in fs, and also implement plugins for different editors which won't need a watcher then.

I decided to go for sending the whole code in POST request instead of just filepath because many time I want to try some code which isn't saved in any file yet, for such cases, editor plugins can simply send the code with a random `filepath` and push it to the app.

The POST request to `/_hotloader` should look like this:

POST http://127.0.0.1:3000/_hotloader
Content-Type: application/json

{
  "path": "/test",         //absolute path of the file, used for identificationn
  "type":  "js",           //type of the file, as needed LiveUpdate's fileHandler
  "oldContent": "tested",  //oldContent, required for `js` code, not for HTML
  "content": "alert('yo')" //new content of the file
}