Make `oldContent` a week dependency
It might not be possible for traditional editors (even for a dedicated watcher) to provide `oldContent`. So now if `oldContent` is not present, we save the present content given for the file, and refresh the page for this one time. For any future pushes, we consider already present content as `oldContent` and hot-push the code.