Add ability to "prepare" files for hotpush
So we always need `oldContent` for hot-pushing javascript. When we don't have oldContent and the pusher don't provide it either, we refresh the page instead.

This commit adds the ability to "prepare" files for hot push. I.e a pusher can "prePush" a file before any changes happen in the file. This `prePush` shall not refresh the page or try to hot-push, but simply store the contents of the file to be used as `oldContent` for future hot-pushing.

A pusher simply need to have `prePush: true` in the POST request to `/_hotloader` for this.