Add watcher for watching for new changes and hot-pushing them
The problem with running a watcher inside meteor app is that then the watcher is as good as the meteor's refresh speed. i.e the new changes are pushed after meteor is done refreshing the client.

I need to figure a way to disable this watcher if:
- External watcher is installed
- Some editor plugin does the pushing job

Detecting the editor plugin won't be possible I believe. So, I think I should make it configurable to allow internal watcher. i.e make it opt in. I will now implement a configuration for `HotLoader` like:

`HotLoader.enableWatch()` or may be `HotLoader.configure({watch: true})`